Move pkgPrompt() to cli.go
